Update command filesystem
